requirements.txt
django
psycopg2
djangorestframework
django-filter
djangorestframework-simplejwt
django-countries